backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diff-match-patch==20241021
Django==5.2.3
django-cors-headers==4.7.0
django-import-export==4.3.9
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dj-database-url==1.3.0
et-xmlfile==1.1.0
fastapi==0.115.14
Flask==3.1.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.2
tablib==3.8.0
trio==0.27.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.35.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.0
